requirements.txt
fastapi==0.115.2 
uvicorn==0.32.0 
python-dotenv==1.0.1 
langchain==0.3.4 
langchain-community==0.3.3 
langchain-chroma==0.1.4 
langchain-huggingface==0.1.0
langchain-redis 
chromadb==0.5.11 
sentence-transformers==3.2.1 
PyPDF2==3.0.1 
pdfplumber==0.11.4
docx2txt==0.8 
pandas==2.2.3 
google-generativeai==0.8.3 
pydantic==2.9.2 
python-multipart==0.0.12
python-docx==1.1.2
structlog==22.1.0
tiktoken==0.8.0 
redis==5.0.0
redisvl

